handle invalid remove index. (index is not integer)

+++ src/main/java/seedu/duke/commands/RemoveCommand.java
@@ -5,32 +5,45 @@ import seedu.duke.exceptions.IllegalValueException;
 import seedu.duke.food.FoodList;
 
 public class RemoveCommand extends Command {
+    public String index;
     public static final String COMMAND_WORD = "remove";
     public static final String BLANK_INDEX_MESSAGE_1 = "Please enter an integer to remove a food item.\n" +
                                                         "You now have ";
     public static final String BLANK_INDEX_MESSAGE_2 = " food items in your list.";
     public static final String INCORRECT_INDEX_MESSAGE_1 = "Incorrect value entered.\n" +
                                                             "Your now have ";
-
     public static final String INVALID_INPUT_MESSAGE = "Please use a reasonable value :<";
-    public String index;
 
     public RemoveCommand(String index) {
         this.index = index;
     }
 
-    public CommandResult execute (FoodList foodlist) throws DukeException {
-
-        if(!isNumberValid(index)){
-            return new CommandResult(INVALID_INPUT_MESSAGE);
+    public boolean isNumeric (){
+        try {
+            Double.parseDouble(index);
+        } catch (NumberFormatException e){
+            return false;
         }
+        return true;
+    }
+
+    public CommandResult execute (FoodList foodlist) throws DukeException {
 
         if (index.isBlank()) {
             String BLANK_INDEX_MESSAGE = BLANK_INDEX_MESSAGE_1 + foodlist.getNumberOfFood() + BLANK_INDEX_MESSAGE_2;
             throw new IllegalValueException(BLANK_INDEX_MESSAGE);
         }
-        int deleteItem = Integer.parseInt(index.trim()) - 1;
 
+        if(!isNumeric()){
+            String BLANK_INDEX_MESSAGE = BLANK_INDEX_MESSAGE_1 + foodlist.getNumberOfFood() + BLANK_INDEX_MESSAGE_2;
+            return new CommandResult(BLANK_INDEX_MESSAGE);
+        }
+
+        if(!isNumberValid(index)){
+            return new CommandResult(INVALID_INPUT_MESSAGE);
+        }
+
+        int deleteItem = Integer.parseInt(index.trim()) - 1;
         if (deleteItem >= foodlist.getNumberOfFood() || deleteItem < 0) {
             String INDEX_MESSAGE = INCORRECT_INDEX_MESSAGE_1 + foodlist.getNumberOfFood() + BLANK_INDEX_MESSAGE_2;
             throw new IllegalValueException(INDEX_MESSAGE);
